requirements.txt
# UI & Dashboard
streamlit>=1.30.0

# Core AI Framework
llama-index>=0.10.0
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai

# Storage & Tools
qdrant-client
llama-index-vector-stores-qdrant
llama-index-tools-tavily-research
tavily-python

# Data Processing
llama-parse
pydantic>=2.0.0
python-dotenv
asyncio